+++ st2common/st2common/models/system/actionchain.py
@@ -19,8 +19,6 @@ import string
 from st2common.util import schema as util_schema
 from st2common.models.api.notification import NotificationSubSchemaAPI
 
-VALIDATOR = util_schema.get_validator(assign_property_default=False)
-
 
 class Node(object):
 
@@ -122,7 +120,8 @@ class ActionChain(object):
     }
 
     def __init__(self, **kw):
-        VALIDATOR(self.schema).validate(kw)
+        util_schema.validate(instance=kw, schema=self.schema, cls=util_schema.CustomValidator,
+                             use_default=False)
 
         for prop in six.iterkeys(self.schema.get('properties', [])):
             value = kw.get(prop, None)