+++ st2common/st2common/validators/api/reactor.py
@@ -25,7 +25,6 @@ __all__ = [
     'validate_trigger_parameters'
 ]
 
-VALIDATOR = util_schema.get_validator(assign_property_default=False)
 allowed_operators = criteria_operators.get_allowed_operators()
 
 
@@ -47,27 +46,27 @@ def validate_criteria(criteria):
                                            'for operator ' + operator)
 
 
-def validate_trigger_parameters(trigger_db):
+def validate_trigger_parameters(trigger_type_ref, parameters):
     """
-    This function validates parameters for system triggers (e.g. timers).
+    This function validates parameters for system triggers (e.g. webhook and timers).
 
     Note: Eventually we should also validate parameters for user defined triggers which correctly
     specify JSON schema for the parameters.
 
-    :param trigger_db: Trigger DB object.
-    :type trigger_db: :class:`TriggerDB`
-    """
-    if not trigger_db:
-        return None
+    :param trigger_type_ref: Reference of a trigger type.
+    :type trigger_type_ref: ``str``
 
-    trigger_type_ref = trigger_db.type
-    parameters = trigger_db.parameters
+    :param parameters: Trigger parameters.
+    :type parameters: ``dict``
+    """
+    if not trigger_type_ref:
+        return None
 
     if trigger_type_ref not in SYSTEM_TRIGGER_TYPES:
         # Not a system trigger, skip validation for now
         return None
 
     parameters_schema = SYSTEM_TRIGGER_TYPES[trigger_type_ref]['parameters_schema']
-    VALIDATOR(parameters_schema).validate(parameters)
-
-    return True
+    cleaned = util_schema.validate(instance=parameters, schema=parameters_schema,
+                                   cls=util_schema.CustomValidator, use_default=True)
+    return cleaned